requirements.txt
Django==1.8.5
defusedxml==0.4.1
django-admin-bootstrapped==2.5.7
django-allauth==0.24.1
django-appconf==1.0.1
django-bootstrap3==6.2.2
django-filter==0.11.0
django-leaflet==0.16.0
django-mptt==0.8.0
et-xmlfile==1.0.1
jdcal==1.2
oauthlib==1.0.3
openpyxl==2.3.0
psycopg2==2.6.1
python3-openid==3.0.9
pytz==2015.7
requests==2.8.1
requests-oauthlib==0.5.0
six==1.10.0
